+++ xarray/core/utils.py
@@ -549,6 +549,11 @@ class ReprObject:
     def __hash__(self) -> int:
         return hash((ReprObject, self._value))
 
+    def __dask_tokenize__(self):
+        from dask.base import normalize_token
+
+        return normalize_token((type(self), self._value))
+
 
 @contextlib.contextmanager
 def close_on_error(f):